Fix flaky DomainSocket tests
Motivation:

`DomainSocketServerTest` frequently fails to delete the domain socket file used for
testing.
https://github.com/line/armeria/actions/runs/5816889007/job/15770694188?pr=5108#step:7:1746
```
DomainSocketServerTest > executionError FAILED
    java.io.UncheckedIOException: java.nio.file.NoSuchFileException: /tmp/armeria17713324805503526473/test.sock
        at java.base/java.nio.file.FileTreeIterator.fetchNextIfNeeded(FileTreeIterator.java:87)
        at java.base/java.nio.file.FileTreeIterator.hasNext(FileTreeIterator.java:103)
        at java.base/java.util.Iterator.forEachRemaining(Iterator.java:132)
        at java.base/java.util.Spliterators$IteratorSpliterator.forEachRemaining(Spliterators.java:1921)
        at java.base/java.util.stream.AbstractPipeline.copyInto(AbstractPipeline.java:509)
        at java.base/java.util.stream.AbstractPipeline.wrapAndCopyInto(AbstractPipeline.java:499)
        at java.base/java.util.stream.ForEachOps$ForEachOp.evaluateSequential(ForEachOps.java:150)
        at java.base/java.util.stream.ForEachOps$ForEachOp$OfRef.evaluateSequential(ForEachOps.java:173)
        at java.base/java.util.stream.AbstractPipeline.evaluate(AbstractPipeline.java:234)
        at java.base/java.util.stream.ReferencePipeline.forEach(ReferencePipeline.java:596)
        at com.linecorp.armeria.internal.testing.TemporaryFolder.delete(TemporaryFolder.java:82)
        at com.linecorp.armeria.internal.testing.TemporaryFolderExtension.after(TemporaryFolderExtension.java:60)
        at com.linecorp.armeria.testing.junit5.common.AbstractAllOrEachExtension.afterAll(AbstractAllOrEachExtension.java:63)
        at java.base/java.util.ArrayList.forEach(ArrayList.java:1511)

        Caused by:
        java.nio.file.NoSuchFileException: /tmp/armeria17713324805503526473/test.sock
            at java.base/sun.nio.fs.UnixException.translateToIOException(UnixException.java:92)
```

After checking the Netty code, I found out a race condition that Netty's
`ServerDomainSocketChannel` also tries to delete the domain socket file
when it is being closed.
https://github.com/netty/netty/blob/9eb029d630b269e907aaa4a19bc86753d70e399c/transport-classes-epoll/src/main/java/io/netty/channel/epoll/EpollServerDomainSocketChannel.java#L79-L81

Modifications:

- Do not automatically delete domain socket files used for testing and
  delegate Netty to delete them

Result:

- Stabilize CI build results
- Fixes #4959

+++ testing-internal/src/main/java/com/linecorp/armeria/internal/testing/TemporaryFolderExtension.java
@@ -45,9 +45,15 @@ public class TemporaryFolderExtension extends AbstractAllOrEachExtension {
     // Forked from https://github.com/line/centraldogma/blob/e5a9c1cf402b7ea59fddb56aae40ebdd1502213e/testing-internal/src/main/java/com/linecorp/centraldogma/testing/internal/TemporaryFolderExtension.java
 
     private final TemporaryFolder delegate;
+    private final boolean autoDelete;
 
     public TemporaryFolderExtension() {
+        this(true);
+    }
+
+    public TemporaryFolderExtension(boolean autoDelete) {
         delegate = new TemporaryFolder();
+        this.autoDelete = autoDelete;
     }
 
     @Override
@@ -57,7 +63,9 @@ public class TemporaryFolderExtension extends AbstractAllOrEachExtension {
 
     @Override
     public void after(ExtensionContext context) throws Exception {
-        delegate.delete();
+        if (autoDelete) {
+            delegate.delete();
+        }
     }
 
     public void create() throws IOException {
